Add files via upload
A roster featuring 219 users from the /r/FBB Discord. Users are organized in OVR by the server's leaderboard rankings (as of February 17th).

Everything is randomly generated. The only things fixed are ratings and names. This means that Ages/Faces/Contracts are randomly generated when loading the file.

When you load into the league, start a Fantasy Draft because everyone will start on Boston.